Moved gateway-dependent classes into Gateway namespace
