+++ src/Gateway/InvoiceItem.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Laravel\Cashier;
+namespace Laravel\Cashier\Gateway;
 
 use Carbon\Carbon;
+use Laravel\Cashier\Cashier;
 
 class InvoiceItem
 {
